Fix disclosure toggling within steps, combine examples, update kit
