+++ src/helpers.ts
@@ -7,23 +7,22 @@ export const joinListBackwards = (
   tr: Transaction,
   nodeType: string,
 ): boolean => {
-  const disclosures = findParentNode((node) => node.type.name === nodeType)(
+  const parent = findParentNode((node) => node.type.name === nodeType)(
     tr.selection,
   );
-  if (!disclosures) return true;
+  if (!parent) return true;
 
   const before = tr.doc
-    .resolve(Math.max(0, disclosures.pos - 1))
-    .before(disclosures.depth);
+    .resolve(Math.max(0, parent.pos - 1))
+    .before(parent.depth);
   if (before === undefined) return true;
 
   const nodeBefore = tr.doc.nodeAt(before);
   const canJoinBackwards =
-    disclosures.node.type === nodeBefore?.type &&
-    canJoin(tr.doc, disclosures.pos);
+    parent.node.type === nodeBefore?.type && canJoin(tr.doc, parent.pos);
   if (!canJoinBackwards) return true;
 
-  tr.join(disclosures.pos);
+  tr.join(parent.pos);
   return true;
 };
 
@@ -31,17 +30,17 @@ export const joinListForwards = (
   tr: Transaction,
   nodeType: string,
 ): boolean => {
-  const disclosures = findParentNode((node) => node.type.name === nodeType)(
+  const parent = findParentNode((node) => node.type.name === nodeType)(
     tr.selection,
   );
-  if (!disclosures) return true;
+  if (!parent) return true;
 
-  const after = tr.doc.resolve(disclosures.start).after(disclosures.depth);
+  const after = tr.doc.resolve(parent.start).after(parent.depth);
   if (after === undefined) return true;
 
   const nodeAfter = tr.doc.nodeAt(after);
   const canJoinForwards =
-    disclosures.node.type === nodeAfter?.type && canJoin(tr.doc, after);
+    parent.node.type === nodeAfter?.type && canJoin(tr.doc, after);
   if (!canJoinForwards) return true;
 
   tr.join(after);
